[client] Add useCreateJob hook for creating a job
The useCreateJob hook is added to the GraphQL hooks module. This hook utilizes the useMutation hook from the @apollo/client package to handle the mutation for creating a job. The hook takes in a title and description as parameters and returns a createJob function and a loading state. When the createJob function is called, it triggers the mutate function and creates a new job with the provided title and description. This function also updates the cache with the newly created job. The loading state indicates whether the mutation is currently in progress.

[client] Update CreateJobPage to use useCreateJob hook

The CreateJobPage component is modified to use the useCreateJob hook instead of the createJob function from the queries module. The useCreateJob hook is used to create a new job with the provided title and description when the form is submitted. The loading state is used to disable the submit button while the mutation is in progress.

[queries] Remove createJob function

The createJob function, which was previously used to create a new job, is no longer needed and has been removed from the queries module.