refactor(server): Add DataLoader for companyLoader
This commit refactors the `createCompanyLoader` function in `server/db/companies.js` to return an instance of DataLoader. This DataLoader is used to batch and cache requests for company data based on their IDs. The function now accepts an array of IDs and retrieves the corresponding companies from the database using the `getCompanyTable().select().whereIn('id', ids)` query. The results are then mapped to the corresponding IDs and returned. This change improves the efficiency of fetching company data by reducing the number of database queries.