feat: Modify getJobs function and add limit and offset parameters
This commit modifies the `getJobs` function in the `jobs.js` file. It adds `limit` and `offset` parameters to the function to allow for pagination when retrieving jobs. The function now returns jobs sorted by `createdAt` in descending order. If `limit` is provided, the query will limit the number of jobs returned, and if `offset` is provided, the query will skip that many jobs. This change enables more control over the number of job results and allows for efficient pagination.