feat: Add ability to count total number of jobs
This commit adds a new function `countJobs` to the `jobs.js` module. This function retrieves the total number of jobs from the database. It is used in the `jobs` query resolver in `resolvers.js` to return a `JobSubList` with both the list of jobs and the total count of jobs.