Refactor JobList rendering and add PaginationBar to HomePage
The modifications in this commit refactor the rendering of JobList in HomePage.js by adding a PaginationBar component. Previously, the page was displaying a set number of jobs per page without any pagination functionality. With the addition of PaginationBar, the jobs are now displayed with pagination support. The total number of jobs is divided into pages, and the user can navigate between them using the "Previous" and "Next" buttons. The currentPage state is updated accordingly. This change improves the user experience by allowing them to view jobs in a more organized and manageable manner.